Fix: Type comparison warnings (-Wtype-limits)
Fixed two instances where comparisons were always true or false due to
limited range of unsigned 8-bit types.

Changes:

1. MenuItems.c:799 - Removed always-false comparison
   PROBLEM: resName[0] is a char (max 255), so `resName[0] > 255` is
   always false.
   FIX: Removed redundant check. Pascal string length byte already limited
   to 255 by type.

2. MenuResources.c:207 - Changed itemOffset type to allow proper bounds checking
   PROBLEM: itemOffset was uint8_t (max 255), so `itemOffset < 512` was
   always true.
   FIX: Changed itemOffset from uint8_t to int, allowing proper comparison
   against resource size limit of 512 bytes and preventing uint8_t overflow
   in line 272 addition.

Both fixes maintain intended behavior while eliminating warnings and
correcting potential overflow issues.

+++ src/MenuManager/MenuResources.c
@@ -202,7 +202,7 @@ MenuHandle ParseMENUResource(Handle resourceHandle)
 
     /* Parse menu items starting after title */
     uint8_t itemCount = data[0x0B + titleLen];
-    uint8_t itemOffset = 0x0C + titleLen;  /* Start of first item */
+    int itemOffset = 0x0C + titleLen;  /* Start of first item (int for bounds checking) */
 
     for (int i = 0; i < itemCount && itemOffset < 512; i++) {
         /* Each item is: length byte + text */
